Add PIDUtils and related classes
An abstraction of several PID controller related functionalities, the primary goal of which is to reduce the highly repetitive and tedious boilerplate associated with working with PID controllers and parameters.

This is done mainly by moving individual parameters into an iterable data structure. Preferably, the configuration would be data-driven as well but I'll leave that be for now....

Examples of the usage and extensibility of these classes can be found in
https://gist.github.com/Alexander3847575/c5ad192a3d2df8a6288e3d86e7ccaa47

(diff recommended)